bpf_progs: use unsigned instead of int
This gets rid of a bunch of warnings (enabled by -Wextra)
about signed vs unsigned comparisons.

For the for loops, the compiler really should be smarter...
the so-called 'unsigned' values are *small* compile time
constants like 20 / 2 = 10 or 40 / 2 = 20.

For the 'int len' argument, it's always going to be a positive
value anyway (negative values or zero are meaningless)

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I57fae9add1d2aea7f47ed6a897f1b0f3fb874113