NetBpfLoad: remove support for 'size_of_bpf_map_def'
Both the loader and all the bpf .o's are in the same apex, they're thus
built from the same source and have the same struct bpf_map_def
definition, and this is no longer required.

We do still need to populate the field in the .o's though,
since offload.o and test.o are still read by the Android S
platform bpfloader.

Afterwards:
  git grep 'DEFAULT_SIZEOF_BPF_MAP_DEF|sizeOfBpfMapDef'
is empty

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I152151e3713b1002f603f298dae060decb4c19fe