Enable secondary_user_on_secondary_display for CtsNetTestCasesInternetPermission
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 364636108
Test: atest CtsNetTestCasesInternetPermission
Test: atest --user-type secondary_user CtsNetTestCasesInternetPermission
Test: atest --user-type secondary_user_on_secondary_display CtsNetTestCasesInternetPermission
Flag: TEST_ONLY

Change-Id: Ib559c8465846f687a725b85c330f398e81cbe567