Get rid of volatile mIsAborted
This has always been race-y as a socket exception could coincide with a
call to #close(). Since it is safe to run the full code paths for both
"regular" exceptions and exceptions triggered by the socket closing,
mIsAborted is not actually needed. This CL gets rid of it in favor of
simplicity.

Test: TH
Change-Id: Ic597ee917daffd6315c349dc435e5533b686bc2d