Remove last @hide usage in CtsNetTestCases
Inline the last remaining constants, and add a workaround for test APIs
that are missing in module API stubs.

While inlining the constants and reflection usage is far from ideal,
this is effectively what the tests are doing today, and it allows
preventing further usage of hidden API in the tests by making them build
against API stubs (in subsequent change).

Test: atest
Flag: TEST_ONLY
Bug: 410435307
Change-Id: Id2ffc0cb41bc31e4cd049b711a392904910414a3