Move setInterfaceUpState out of try / catch block
setInterfaceUpState does not need to be in the try / catch block as it
doesn't let any exceptions escape. Also move the call to after checking
whether the returned config is null. In practice, this should never
happen; but if it did, we wouldn't end up doing anything anyways, so
bringing up the interface is pointless.

This also annotates config as final and returns early out of the catch
block (because the second log statement is technically incorrect).

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: Ic9233784594ead6779caf6267afad7b518a873ea