[Thread] fix SparseArray hash for ActiveOperationalDataset
The existing ActiveOperationalDataset#hashCode has an issue that it's
depending on the SparseArray#hasCode which doesn't override the
Object#hashCode method. The consequence is that when the
ActiveOperationalDataset is used as a key in a container, there will
be a miss even if two ActiveOperationalDataset objects have the same
content.

This commit fixes this issue by using a dedicated
byteArraySparseArrayHasCode() method which computes the content hashcode
with Arrays.deepHashCode(). Also fixes the issue that mUnknownTlvs is
not counted in hashCode().

Flag: EXEMPT trivial bugfix
Change-Id: I085d3bf0bdf4bb73bb995fc28dd3857437f9743e