Replace shouldTrackInterface() with getTrackingReason()
In order to bring up a local-only NetworkOffer for USB NCM interfaces
(independent of whether they are included in the regex), EthernetTracker
must pass the following (non-exclusive) information to the factory:
1) whether the interface matches the regex.
2) whether the interface is an NCM interface.

If the interface matches the regex, the EthernetNetworkFactory creates a
global offer for it.
If the interface is an NCM interface, the EthernetNetworkFactory creates
a local offer.

This change adds the first step in that direction by adding a
TrackingReason enum and returning it from getTrackingReason(). However,
EthernetTracker continues to ignore NCM interfaces if they are not part
of the regex. This will be handled later, as soon as the reason is piped
all the way to the factory to prevent bringing up a global offer for
such interfaces.

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: I31f38125f2d5cc0072ea34be45f40afb110dfbb6