Add RequestTracker to keep track of both global and local requests
This partially "reverts" a previous commit that moved mRequestIds into
the EthernetNetworkOfferCallback. It turns out, a global object is
needed to properly keep track of both local and global network requests.
Additionally, a Mode enum is added and passed to start() to specify
whether to bring up a global or local network. The mode cannot be
changed without stopping the current network first.

Note that the LocalNetworkOfferCallback does not register its requests
with the RequestTracker yet, meaning Mode.LOCAL is still unused as of
this CL.

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: If40d057a79a99bf68cfc8150fe6d851b3bf7669c