Delete low value log
This log statement was not particularly useful and borderline read
like an error / warning (despite its info level). This CL removes it.

For logs in general, it is most useful to know a) things are happening, or b)
things were prevented from happening due to some error condition. This
fits in neither category.

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: I2ec21d5cf3d0d88feb34af0e71450c30217a63cf